+++ src/Helpers/pluginHelpers.js
@@ -1,4 +1,0 @@
-/**
- * Created by Martin Neundorfer on 06.09.2018.
- * For LABOR.digital
- */